+++ Code/Mantid/Framework/Algorithms/src/Integration.cpp
@@ -66,7 +66,7 @@ void Integration::exec()
   // Get the input workspace
   MatrixWorkspace_const_sptr localworkspace = getProperty("InputWorkspace");
 
-  const int numberOfSpectra = localworkspace->getNumberHistograms();
+  const size_t numberOfSpectra = localworkspace->getNumberHistograms();
 
   // Check 'StartSpectrum' is in range 0-numberOfSpectra
   if ( m_MinSpec > numberOfSpectra )
